+++ Realm/RLMObjectSchema_Private.h
@@ -41,7 +41,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readwrite, assign) Class accessorClass;
 @property (nonatomic, readwrite, assign) Class unmanagedClass;
 
-@property (nonatomic, readwrite, assign) bool hasCustomAuditSerialization;
+@property (nonatomic, readwrite, assign) bool hasCustomEventSerialization;
 
 @property (nonatomic, readwrite, nullable) RLMProperty *primaryKeyProperty;