+++ Realm/RLMRealmConfiguration.mm
@@ -18,7 +18,7 @@
 
 #import "RLMRealmConfiguration_Private.h"
 
-#import "RLMAudit.h"
+#import "RLMEvent.h"
 #import "RLMObjectSchema_Private.hpp"
 #import "RLMRealm_Private.h"
 #import "RLMSchema_Private.hpp"
@@ -389,8 +389,8 @@ static bool isSync(realm::Realm::Config const& config) {
     if (config.sync_config) {
         RLMSetConfigInfoForClientResetCallbacks(*config.sync_config, self);
     }
-    if (_auditConfiguration) {
-        config.audit_config = [_auditConfiguration auditConfigWithRealmConfiguration:self];
+    if (_eventConfiguration) {
+        config.audit_config = [_eventConfiguration auditConfigWithRealmConfiguration:self];
     }
 #endif
     return config;