+++ Realm/RLMRealmConfiguration_Private.h
@@ -18,7 +18,7 @@
 
 #import <Realm/RLMRealmConfiguration.h>
 
-@class RLMSchema, RLMAuditConfiguration;
+@class RLMSchema, RLMEventConfiguration;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -29,7 +29,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readwrite) bool disableFormatUpgrade;
 @property (nonatomic, copy, nullable) RLMSchema *customSchema;
 @property (nonatomic, copy) NSString *pathOnDisk;
-@property (nonatomic, retain, nullable) RLMAuditConfiguration *auditConfiguration;
+@property (nonatomic, retain, nullable) RLMEventConfiguration *eventConfiguration;
 
 // Get the default confiugration without copying it
 + (RLMRealmConfiguration *)rawDefaultConfiguration;